infra: the CUDA batch arena, and a repository that can reproduce its hashes
The CPU baseline import had no GPU path and no way to run more than one
statepoint per process.  This lands the machinery both the 40x ledger and the
GA pipeline are measured against:

  * src/CudaBICGBackend.{cu,h} + Stub.cpp -- two-group CMFD BiCGSTAB device
    kernels, BatchCore, CudaBatchArena and CUDA Graph capture.  The arena is
    deliberately *opportunistic*, not lock-step: a batch is "everyone who
    arrived while the previous batch was on the device", which is only sound
    because a slot's answer does not depend on its fellow passengers
    (CudaBICGBackend.cu:1240-1262).  Registering under the arena lock first
    collapsed mean width to 2.1/8; registration is lock-free now.
  * MirroredUpload elision (:581-607): `cc` and `phi` skip the DMA when they
    are bit-identical to the host shadow.  M1 skips 6,520 of 8,156 H2D copies
    (79.9%) -- GPU_RASBERRY_QUARTER_GA_OPTIMIZATION_FINAL_20260820_KO.md §4.
  * RASBERY_BATCH_WAIT_US defaults to 0 on purpose.  Linger buys nothing while
    the device holds 4.7% of the wall; it becomes worth paying for only when
    more of the step is device-resident.
  * Scheduler/Geometry/BICG* carry the batch-mode statepoint plumbing and
    rasbery_omp_gate.
  * src/Timer.h is deleted; nothing referenced it.

.gitattributes is the load-bearing part of this commit.  It was
`* text=auto eol=lf`, which stores and checks out LF.  Every SHA-256 the
ledger pins is a hash of the *CRLF* working-tree bytes --

    src/XSSet.cpp  0305496D2613C8D02668076F243CD856A61131B62B70978A96AC65B82C4D9FD5

-- so a clone on the Linux build host would have compiled different bytes and
missed every pinned hash.  The authoritative source paths are now `-text`;
vendored trees keep the old default so this does not rewrite files nobody
hashes.

.gitignore: `*.pngbuild/` was one line, a typo fusing `*.png` and `build/`
into a pattern that matched nothing.  test/ (3.9 GB of CHIFFON libraries, HGC
tables and recorded baselines) is excluded the same way the consolidated
snapshot excludes it, with C++ contract tests re-included.

Receipts: GPU_RASBERRY_40X_OPTIMIZATION_LEDGER_20260820.md,
combined_patch_receipt_20260820.md (topology guard + adaptive rendezvous),
ANALYSIS_STRUCTURE_20260820.md §2.2-§2.3.

+++ src/pch.h
@@ -6,6 +6,8 @@
 #include <array>
 #include <chrono>
 #include <cmath>
+#include <cstdio>
+#include <cstdlib>
 #include <cstring>
 #include <iostream>
 #include <map>
@@ -75,13 +77,21 @@ static const int NG2 = 2;
 
 static const int NPTM = 3;
 
-static const float HAW   = 1.0079;
-static const float OAW   = 15.994915;
-static const float H2OAW = 18.010715;
+// Atomic/molecular weights and Avogadro's number for the number-density relation
+// N = rho * AVOG / AW.  Held as float they carried only ~7 significant digits into what is
+// otherwise an all-double chain, so any future caller would inherit a ~1e-7 relative bias on
+// every boron / moderator number density -- systematic, not round-off, because the same
+// rounded literal is reused for every node.  Verified: as of this commit none of the six has
+// a call site anywhere in the tree, so widening them changes no result today.  They are kept
+// (rather than deleted) because they are the declared constants for a relation the solver
+// still needs; this makes them safe to pick up.
+static const double HAW   = 1.0079;
+static const double OAW   = 15.994915;
+static const double H2OAW = 18.010715;
 
-static const float AVOG  = 0.6022045;
-static const float B10AW = 10.012937;
-static const float B11AW = 11.009305;
+static const double AVOG  = 0.6022045;
+static const double B10AW = 10.012937;
+static const double B11AW = 11.009305;
 
 static const int TF_POINT = 20;
